refactor(settings): inline BaseScoreTile
Remove BaseScoreTile parent. DimerTile and ReplicationTile now
directly inherit ft.ExpansionTile. init logic duplicated.
ScoreTable retained as standalone component.
