+++ src/amplifyp/gui/views/settings/__init__.py
@@ -17,10 +17,7 @@
 
 from amplifyp.gui.views.settings.appearance_tile import AppearanceTile
 from amplifyp.gui.views.settings.backup_tile import BackupTile
-from amplifyp.gui.views.settings.base_score_tile import (
-    BaseScoreTile,
-    ScoreTable,
-)
+from amplifyp.gui.views.settings.base_score_tile import ScoreTable
 from amplifyp.gui.views.settings.dimer_tile import DimerTile
 from amplifyp.gui.views.settings.replication_tile import ReplicationTile
 from amplifyp.gui.views.settings.settings_view import SettingsView
@@ -29,7 +26,6 @@ from amplifyp.gui.views.settings.tm_tile import TmTile
 __all__ = [
     "AppearanceTile",
     "BackupTile",
-    "BaseScoreTile",
     "DimerTile",
     "ReplicationTile",
     "ScoreTable",
