+++ src/amplifyp/gui/views/settings/dimer_tile.py
@@ -22,10 +22,10 @@ import flet as ft
 
 from amplifyp.gui.colours import GUIColours
 from amplifyp.gui.settings import GUISettings
-from amplifyp.gui.views.settings.base_score_tile import BaseScoreTile
+from amplifyp.gui.views.settings.base_score_tile import ScoreTable
 
 
-class DimerTile(BaseScoreTile):
+class DimerTile(ft.ExpansionTile):  # type: ignore[misc]
     """Expansion tile for Dimer settings."""
 
     def __init__(
@@ -50,7 +50,12 @@ class DimerTile(BaseScoreTile):
             font_size_micro: Micro font size for small labels.
             font_size_table_header: Font size for table header cells.
         """
+        self.settings = settings
+        self.settings_map = settings_map
+        self.on_change_handler = on_change_handler
+
         from amplifyp.dna import Nucleotides
+        from amplifyp.gui.utils.ui import initialise_score_fields
 
         self.set_pd_min_overlap = ft.TextField(
             label="Min Overlap",
@@ -68,23 +73,68 @@ class DimerTile(BaseScoreTile):
         settings_map["pd_min_overlap"] = self.set_pd_min_overlap
         settings_map["pd_threshold"] = self.set_pd_threshold
 
-        super().__init__(
-            settings=settings,
-            settings_map=settings_map,
-            on_change_handler=on_change_handler,
-            header_size=header_size,
-            font_size_default=font_size_default,
-            font_size_micro=font_size_micro,
-            font_size_table_header=font_size_table_header,
-            title="Primer Dimer Settings",
-            score_table_label="Primer Dimer Weights",
-            score_table_prefix="pd_score",
+        initialise_score_fields(
+            settings_map=self.settings_map,
+            prefix="pd_score",
+            row_headers=list(Nucleotides.PRIMER),
+            col_headers=list(Nucleotides.PRIMER),
+            on_change_handler=self.on_change_handler,
+            font_size=font_size_default,
+        )
+
+        self.score_table = ScoreTable(
+            label="Primer Dimer Weights",
             row_headers=list(Nucleotides.PRIMER),
             col_headers=list(Nucleotides.PRIMER),
             row_label="Primer",
             col_label="Primer",
-            parameter_controls=[
-                self.set_pd_min_overlap,
-                self.set_pd_threshold,
+            prefix="pd_score",
+            settings_map=self.settings_map,
+            font_size_default=font_size_default,
+            font_size_micro=font_size_micro,
+            font_size_table_header=font_size_table_header,
+        )
+
+        super().__init__(
+            title=ft.Text(
+                "Primer Dimer Settings",
+                weight=ft.FontWeight.BOLD,
+                size=header_size,
+            ),
+            expanded_cross_axis_alignment=ft.CrossAxisAlignment.STRETCH,
+            controls=[
+                ft.Container(
+                    content=ft.Column(
+                        [
+                            self.score_table,
+                            ft.Container(height=10),
+                            ft.Container(
+                                content=ft.Column(
+                                    [
+                                        ft.Text(
+                                            "Parameters",
+                                            weight=ft.FontWeight.BOLD,
+                                            size=self.settings.get(
+                                                "font_size_default", 14
+                                            ),
+                                        ),
+                                        self.set_pd_min_overlap,
+                                        self.set_pd_threshold,
+                                    ],
+                                    spacing=15,
+                                    horizontal_alignment=ft.CrossAxisAlignment.STRETCH,
+                                ),
+                                width=500,
+                            ),
+                        ],
+                        horizontal_alignment=ft.CrossAxisAlignment.CENTER,
+                        spacing=10,
+                    ),
+                    padding=ft.Padding(0, 20, 0, 10),
+                )
             ],
         )
+
+    def update_ui(self) -> None:
+        """Update Flet UI controls to match theme/settings."""
+        self.score_table.table.heading_row_color = GUIColours.INFO_HEADER_BG
