+++ src/amplifyp/gui/views/settings/replication_tile.py
@@ -22,10 +22,10 @@ import flet as ft
 
 from amplifyp.gui.colours import GUIColours
 from amplifyp.gui.settings import GUISettings
-from amplifyp.gui.views.settings.base_score_tile import BaseScoreTile
+from amplifyp.gui.views.settings.base_score_tile import ScoreTable
 
 
-class ReplicationTile(BaseScoreTile):
+class ReplicationTile(ft.ExpansionTile):  # type: ignore[misc]
     """Expansion tile for Origin of Replication settings."""
 
     def __init__(
@@ -50,7 +50,15 @@ class ReplicationTile(BaseScoreTile):
             font_size_micro (int): Micro font size for small text.
             font_size_table_header (int): Font size for table headers.
         """
+        self.settings = settings
+        self.settings_map = settings_map
+        self.on_change_handler = on_change_handler
+
         from amplifyp.dna import Nucleotides
+        from amplifyp.gui.utils.ui import (
+            BorderedCheckbox,
+            initialise_score_fields,
+        )
 
         self.set_primability_cutoff = ft.TextField(
             label="Primability Cutoff",
@@ -64,7 +72,6 @@ class ReplicationTile(BaseScoreTile):
             on_change=on_change_handler,
             border_color=GUIColours.OUTLINE,
         )
-        from amplifyp.gui.utils.ui import BorderedCheckbox
 
         self.set_amp4_compat = BorderedCheckbox(
             label="Amplify4 Compatibility Mode",
@@ -83,25 +90,70 @@ class ReplicationTile(BaseScoreTile):
 
         col_headers = [c for c in Nucleotides.TEMPLATE if c != Nucleotides.GAP]
 
-        super().__init__(
-            settings=settings,
-            settings_map=settings_map,
-            on_change_handler=on_change_handler,
-            header_size=header_size,
-            font_size_default=font_size_default,
-            font_size_micro=font_size_micro,
-            font_size_table_header=font_size_table_header,
-            title="Origin of Replication Settings",
-            score_table_label="Base Pair Weights",
-            score_table_prefix="bp_score",
+        initialise_score_fields(
+            settings_map=self.settings_map,
+            prefix="bp_score",
+            row_headers=list(Nucleotides.PRIMER),
+            col_headers=col_headers,
+            on_change_handler=self.on_change_handler,
+            font_size=font_size_default,
+        )
+
+        self.score_table = ScoreTable(
+            label="Base Pair Weights",
             row_headers=list(Nucleotides.PRIMER),
             col_headers=col_headers,
             row_label="Primer",
             col_label="Template",
-            parameter_controls=[
-                self.set_primability_cutoff,
-                self.set_stability_cutoff,
-                self.set_amp4_compat,
-                self.set_improved_visualisation,
+            prefix="bp_score",
+            settings_map=self.settings_map,
+            font_size_default=font_size_default,
+            font_size_micro=font_size_micro,
+            font_size_table_header=font_size_table_header,
+        )
+
+        super().__init__(
+            title=ft.Text(
+                "Origin of Replication Settings",
+                weight=ft.FontWeight.BOLD,
+                size=header_size,
+            ),
+            expanded_cross_axis_alignment=ft.CrossAxisAlignment.STRETCH,
+            controls=[
+                ft.Container(
+                    content=ft.Column(
+                        [
+                            self.score_table,
+                            ft.Container(height=10),
+                            ft.Container(
+                                content=ft.Column(
+                                    [
+                                        ft.Text(
+                                            "Parameters",
+                                            weight=ft.FontWeight.BOLD,
+                                            size=self.settings.get(
+                                                "font_size_default", 14
+                                            ),
+                                        ),
+                                        self.set_primability_cutoff,
+                                        self.set_stability_cutoff,
+                                        self.set_amp4_compat,
+                                        self.set_improved_visualisation,
+                                    ],
+                                    spacing=15,
+                                    horizontal_alignment=ft.CrossAxisAlignment.STRETCH,
+                                ),
+                                width=500,
+                            ),
+                        ],
+                        horizontal_alignment=ft.CrossAxisAlignment.CENTER,
+                        spacing=10,
+                    ),
+                    padding=ft.Padding(0, 20, 0, 10),
+                )
             ],
         )
+
+    def update_ui(self) -> None:
+        """Update Flet UI controls to match theme/settings."""
+        self.score_table.table.heading_row_color = GUIColours.INFO_HEADER_BG
